cdk/apps/ica_credentials/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
attrs==21.2.0
    # via
    #   cattrs
    #   jsii
aws-cdk.assets==1.120.0
    # via
    #   aws-cdk.aws-ecr-assets
    #   aws-cdk.aws-s3-assets
aws-cdk.aws-applicationautoscaling==1.120.0
    # via aws-cdk.aws-lambda
aws-cdk.aws-autoscaling-common==1.120.0
    # via aws-cdk.aws-applicationautoscaling
aws-cdk.aws-cloudwatch==1.120.0
    # via
    #   aws-cdk.aws-applicationautoscaling
    #   aws-cdk.aws-ec2
    #   aws-cdk.aws-lambda
    #   aws-cdk.aws-logs
    #   aws-cdk.aws-sqs
aws-cdk.aws-codeguruprofiler==1.120.0
    # via aws-cdk.aws-lambda
aws-cdk.aws-ec2==1.120.0
    # via
    #   aws-cdk.aws-efs
    #   aws-cdk.aws-lambda
    #   aws-cdk.aws-secretsmanager
aws-cdk.aws-ecr==1.120.0
    # via
    #   aws-cdk.aws-ecr-assets
    #   aws-cdk.aws-lambda
aws-cdk.aws-ecr-assets==1.120.0
    # via aws-cdk.aws-lambda
aws-cdk.aws-efs==1.120.0
    # via aws-cdk.aws-lambda
aws-cdk.aws-events==1.120.0
    # via
    #   aws-cdk.aws-ecr
    #   aws-cdk.aws-lambda
    #   aws-cdk.aws-s3
aws_cdk.aws_iam==1.120.0
    # via
    #   -r requirements.in
    #   aws-cdk.aws-applicationautoscaling
    #   aws-cdk.aws-autoscaling-common
    #   aws-cdk.aws-cloudwatch
    #   aws-cdk.aws-codeguruprofiler
    #   aws-cdk.aws-ec2
    #   aws-cdk.aws-ecr
    #   aws-cdk.aws-ecr-assets
    #   aws-cdk.aws-efs
    #   aws-cdk.aws-events
    #   aws-cdk.aws-kms
    #   aws-cdk.aws-lambda
    #   aws-cdk.aws-logs
    #   aws-cdk.aws-s3
    #   aws-cdk.aws-s3-assets
    #   aws-cdk.aws-secretsmanager
    #   aws-cdk.aws-sqs
    #   aws-cdk.aws-ssm
aws-cdk.aws-kms==1.120.0
    # via
    #   aws-cdk.aws-ec2
    #   aws-cdk.aws-efs
    #   aws-cdk.aws-lambda
    #   aws-cdk.aws-logs
    #   aws-cdk.aws-s3
    #   aws-cdk.aws-s3-assets
    #   aws-cdk.aws-secretsmanager
    #   aws-cdk.aws-sqs
    #   aws-cdk.aws-ssm
aws_cdk.aws_lambda==1.120.0
    # via
    #   -r requirements.in
    #   aws-cdk.aws-secretsmanager
aws-cdk.aws-logs==1.120.0
    # via
    #   aws-cdk.aws-ec2
    #   aws-cdk.aws-lambda
aws-cdk.aws-s3==1.120.0
    # via
    #   aws-cdk.aws-ec2
    #   aws-cdk.aws-ecr-assets
    #   aws-cdk.aws-lambda
    #   aws-cdk.aws-s3-assets
aws-cdk.aws-s3-assets==1.120.0
    # via
    #   aws-cdk.aws-ec2
    #   aws-cdk.aws-lambda
    #   aws-cdk.aws-logs
aws-cdk.aws-sam==1.120.0
    # via aws-cdk.aws-secretsmanager
aws-cdk.aws-secretsmanager==1.120.0
    # via -r requirements.in
aws-cdk.aws-signer==1.120.0
    # via aws-cdk.aws-lambda
aws-cdk.aws-sqs==1.120.0
    # via aws-cdk.aws-lambda
aws-cdk.aws-ssm==1.120.0
    # via aws-cdk.aws-ec2
aws-cdk.cloud-assembly-schema==1.120.0
    # via
    #   aws-cdk.aws-ec2
    #   aws-cdk.aws-efs
    #   aws-cdk.aws-ssm
    #   aws-cdk.core
    #   aws-cdk.cx-api
aws_cdk.core==1.120.0
    # via
    #   -r requirements.in
    #   aws-cdk.assets
    #   aws-cdk.aws-applicationautoscaling
    #   aws-cdk.aws-autoscaling-common
    #   aws-cdk.aws-cloudwatch
    #   aws-cdk.aws-codeguruprofiler
    #   aws-cdk.aws-ec2
    #   aws-cdk.aws-ecr
    #   aws-cdk.aws-ecr-assets
    #   aws-cdk.aws-efs
    #   aws-cdk.aws-events
    #   aws-cdk.aws-iam
    #   aws-cdk.aws-kms
    #   aws-cdk.aws-lambda
    #   aws-cdk.aws-logs
    #   aws-cdk.aws-s3
    #   aws-cdk.aws-s3-assets
    #   aws-cdk.aws-sam
    #   aws-cdk.aws-secretsmanager
    #   aws-cdk.aws-signer
    #   aws-cdk.aws-sqs
    #   aws-cdk.aws-ssm
aws-cdk.cx-api==1.120.0
    # via
    #   aws-cdk.assets
    #   aws-cdk.aws-ec2
    #   aws-cdk.aws-ecr-assets
    #   aws-cdk.aws-efs
    #   aws-cdk.aws-kms
    #   aws-cdk.aws-lambda
    #   aws-cdk.aws-s3
    #   aws-cdk.aws-s3-assets
    #   aws-cdk.aws-secretsmanager
    #   aws-cdk.core
aws-cdk.region-info==1.120.0
    # via
    #   aws-cdk.aws-ec2
    #   aws-cdk.aws-iam
    #   aws-cdk.aws-lambda
    #   aws-cdk.core
boto3==1.18.32
    # via -r requirements.in
botocore==1.21.32
    # via
    #   boto3
    #   s3transfer
cattrs==1.8.0
    # via jsii
constructs==3.3.138
    # via
    #   aws-cdk.assets
    #   aws-cdk.aws-applicationautoscaling
    #   aws-cdk.aws-autoscaling-common
    #   aws-cdk.aws-cloudwatch
    #   aws-cdk.aws-codeguruprofiler
    #   aws-cdk.aws-ec2
    #   aws-cdk.aws-ecr
    #   aws-cdk.aws-ecr-assets
    #   aws-cdk.aws-efs
    #   aws-cdk.aws-events
    #   aws-cdk.aws-iam
    #   aws-cdk.aws-kms
    #   aws-cdk.aws-lambda
    #   aws-cdk.aws-logs
    #   aws-cdk.aws-s3
    #   aws-cdk.aws-s3-assets
    #   aws-cdk.aws-sam
    #   aws-cdk.aws-secretsmanager
    #   aws-cdk.aws-signer
    #   aws-cdk.aws-sqs
    #   aws-cdk.aws-ssm
    #   aws-cdk.core
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
jsii==1.34.0
    # via
    #   aws-cdk.assets
    #   aws-cdk.aws-applicationautoscaling
    #   aws-cdk.aws-autoscaling-common
    #   aws-cdk.aws-cloudwatch
    #   aws-cdk.aws-codeguruprofiler
    #   aws-cdk.aws-ec2
    #   aws-cdk.aws-ecr
    #   aws-cdk.aws-ecr-assets
    #   aws-cdk.aws-efs
    #   aws-cdk.aws-events
    #   aws-cdk.aws-iam
    #   aws-cdk.aws-kms
    #   aws-cdk.aws-lambda
    #   aws-cdk.aws-logs
    #   aws-cdk.aws-s3
    #   aws-cdk.aws-s3-assets
    #   aws-cdk.aws-sam
    #   aws-cdk.aws-secretsmanager
    #   aws-cdk.aws-signer
    #   aws-cdk.aws-sqs
    #   aws-cdk.aws-ssm
    #   aws-cdk.cloud-assembly-schema
    #   aws-cdk.core
    #   aws-cdk.cx-api
    #   aws-cdk.region-info
    #   constructs
publication==0.0.3
    # via
    #   aws-cdk.assets
    #   aws-cdk.aws-applicationautoscaling
    #   aws-cdk.aws-autoscaling-common
    #   aws-cdk.aws-cloudwatch
    #   aws-cdk.aws-codeguruprofiler
    #   aws-cdk.aws-ec2
    #   aws-cdk.aws-ecr
    #   aws-cdk.aws-ecr-assets
    #   aws-cdk.aws-efs
    #   aws-cdk.aws-events
    #   aws-cdk.aws-iam
    #   aws-cdk.aws-kms
    #   aws-cdk.aws-lambda
    #   aws-cdk.aws-logs
    #   aws-cdk.aws-s3
    #   aws-cdk.aws-s3-assets
    #   aws-cdk.aws-sam
    #   aws-cdk.aws-secretsmanager
    #   aws-cdk.aws-signer
    #   aws-cdk.aws-sqs
    #   aws-cdk.aws-ssm
    #   aws-cdk.cloud-assembly-schema
    #   aws-cdk.core
    #   aws-cdk.cx-api
    #   aws-cdk.region-info
    #   constructs
python-dateutil==2.8.2
    # via
    #   botocore
    #   jsii
s3transfer==0.5.0
    # via boto3
six==1.16.0
    # via python-dateutil
typing-extensions==3.10.0.2
    # via jsii
urllib3==1.26.6
    # via botocore
